+++ dev/metaParse.ts
@@ -1,8 +1,8 @@
 import fs from 'fs';
-import { getMeta } from '../store/getMeta';
+import { getMetaFromUrl } from '../store/getMeta';
 
 async function start() {
-  const message = await getMeta(
+  const message = await getMetaFromUrl(
     'http://replay117.valve.net/570/7468445438_1951738768.meta.bz2'
   );
   // Stats: Original bzip2, 77kb, unzipped, 113kb, parsed JSON 816kb
