+++ store/getMeta.ts
@@ -3,6 +3,9 @@ import axios from 'axios';
 import fs from 'fs';
 import { exec } from 'child_process';
 import { promisify } from 'util';
+import db from './db';
+import redis from './redis';
+import { buildReplayUrl, redisCount } from '../util/utility';
 const execPromise = promisify(exec);
 
 // Get a sample meta file
@@ -14,16 +17,38 @@ const builder = root.loadSync('./proto/dota_match_metadata.proto', {
 });
 const Message = builder.lookupType('CDOTAMatchMetadataFile');
 
-export async function getMeta(url: string) {
+export async function getMeta(matchId: string) {
+  // Check if we have the info in match_gcdata to construct url
+  const saved = await db.raw(
+    'select match_id, cluster, replay_salt from match_gcdata where match_id = ?',
+    [matchId]
+  );
+  const gcdata = saved.rows[0];
+  if (!gcdata) {
+    return null;
+  }
+  const url = buildReplayUrl(gcdata.match_id, gcdata.cluster, gcdata.replay_salt, true);
+  // Parse it from url
+  // This takes about 50ms of CPU time per match
+  const message = await getMetaFromUrl(url);
+  if (message) {
+    // Count the number of meta parses
+    redisCount(redis, 'meta_parse');
+  }
+  // Return the info, it may be null if we failed at any step or meta isn't available
+  return message;
+}
+
+export async function getMetaFromUrl(url: string) {
   try {
-    console.time('download/bunzip');
+    console.time('[METAPARSE]: download/bunzip');
     const { stdout } = await execPromise(
       `curl -L ${url} | bunzip2`,
       //@ts-ignore
       { shell: true, encoding: 'buffer', maxBuffer: 10 * 1024 * 1024 }
     );
-    console.timeEnd('download/bunzip');
-    console.time('metaParse');
+    console.timeEnd('[METAPARSE]: download/bunzip');
+    console.time('[METAPARSE]: parse');
     const message: any = Message.decode(stdout);
     // message.metadata.teams.forEach((team) => {
     //   team.players.forEach((player) => {
@@ -34,7 +59,7 @@ export async function getMeta(url: string) {
     // });
     // This is encrypted in some way and it's not clear how to read it
     delete message.private_metadata;
-    console.timeEnd('metaParse');
+    console.timeEnd('[METAPARSE]: parse');
     return message;
   } catch (e) {
     console.error(e);
